// src/models/images.ts
export enum DefaultImages{
    default = "https://medieval-tavern-api.azurewebsites.net/images/default-user-image.jpeg",
    manElf1 = "https://medieval-tavern-api.azurewebsites.net/images/man-elf-1.jpeg",
    manElf2 = "https://medieval-tavern-api.azurewebsites.net/images/man-elf-2.jpeg",
    manOrc = "https://medieval-tavern-api.azurewebsites.net/images/man-orc.jpeg",
    manTiefling = "https://medieval-tavern-api.azurewebsites.net/images/man-tiefling.jpeg",
    womanElf1 = "https://medieval-tavern-api.azurewebsites.net/images/woman-elf-1.jpeg",
    womanElf2 = "https://medieval-tavern-api.azurewebsites.net/images/woman-elf-2.jpeg",
    womanOrc = "https://medieval-tavern-api.azurewebsites.net/images/woman-orc.jpeg",
    womanTiefling = "https://medieval-tavern-api.azurewebsites.net/images/woman-tiefling.jpeg",
}